requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
openai==1.6.1
google-generativeai==0.3.1
pymongo==4.5.0
flasgger==0.9.7 